✨ (yaldap): Add password hash mecanism
Storing plain-text password is not a good practice at all. So, in order
to improve a bit the security of this project, I add the ability to
store hashed password that can be managed transparently during the LDAP
BIND request.

Currently, yaLDAP only manages argon2, bcrypt, pbkdf2 and scrypt hashed
password.